+++ common/evolvables/knn.h
@@ -6,10 +6,10 @@
 
 namespace KNN {
     template<typename T>
-    std::vector<std::vector<std::pair<unsigned int,float>>> KNN(const T& mass_group, unsigned int K);
+    std::vector<std::vector<std::pair<uint16_t,float>>> KNN(const T& mass_group, uint16_t K);
 
     template<typename T>
-    std::vector<std::vector<std::pair<unsigned int,float>>> KNN_CPU(const T& mass_group, unsigned int K);
+    std::vector<std::vector<std::pair<uint16_t,float>>> KNN_CPU(const T& mass_group, uint16_t K);
 
     /// @brief Computes the K Nearest neighbors for each mass of each mass_list
     /// @tparam T - requires variable masses
@@ -17,7 +17,7 @@ namespace KNN {
     /// @param k 
     /// @return vector of KNNs for each mass in each group
     template<typename T>
-    std::vector<std::vector<std::vector<std::pair<unsigned int,float>>>> Batch(const std::vector<T>& mass_groups, unsigned int K);
+    std::vector<std::vector<std::vector<std::pair<uint16_t,float>>>> Batch(const std::vector<T>& mass_groups, uint16_t K);
 }
 
 #endif
